requirements-minimal.txt
# Core Framework (Essential) - Using latest versions with pre-built wheels
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# HTTP and Scraping (Essential)
httpx>=0.28.0
requests>=2.32.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11

# Storage and Cache (Essential)
redis>=5.2.0
sqlalchemy>=2.0.36

# Translation (Essential)
deep-translator>=1.11.4

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Utilities (Essential)
python-dotenv>=1.0.1
python-multipart>=0.0.12
slowapi>=0.1.9

# Video processing (Optional - for Rutube)
yt-dlp>=2023.12.30
openai-whisper>=20231117
